backend/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
google-api-python-client
transformers
torch